Refactored user vue components
