+++ resources/js/Pages/User/Partials/NavigationTabs.vue
